Get rid of Linear blocks in the RNN
