+++ src/models/neural_networks/base.py
@@ -4,7 +4,6 @@ from pathlib import Path
 import math
 
 import torch
-from torch import nn
 from torch.nn import functional as F
 
 import shap
@@ -230,20 +229,3 @@ class NNBase(ModelBase):
             assert indices is not None, f"Years can't be None if include pred months is True"
             return torch.eye(14)[indices.long()][:, 1:-1]
         return None
-
-
-class LinearBlock(nn.Module):
-    """
-    A linear layer followed by batchnorm, a ReLU activation, and dropout
-    """
-
-    def __init__(self, in_features, out_features, dropout=0.25):
-        super().__init__()
-        self.linear = nn.Linear(in_features=in_features, out_features=out_features, bias=False)
-        self.relu = nn.LeakyReLU(negative_slope=0.1, inplace=True)
-        self.batchnorm = nn.BatchNorm1d(num_features=out_features)
-        self.dropout = nn.Dropout(dropout)
-
-    def forward(self, x):
-        x = self.relu(self.batchnorm(self.linear(x)))
-        return self.dropout(x)
