+++ tests/models/neural_networks/test_rnn.py
@@ -18,13 +18,12 @@ class TestRecurrentNetwork:
         dense_features = [10]
         hidden_size = 128
         rnn_dropout = 0.25
-        dense_dropout = 0.25
         include_pred_month = True
         include_latlons = True
 
         def mocktrain(self):
             self.model = RNN(features_per_month, dense_features, hidden_size,
-                             rnn_dropout, dense_dropout,
+                             rnn_dropout,
                              include_pred_month, include_latlons,
                              experiment='one_month_forecast')
             self.features_per_month = features_per_month
@@ -50,7 +49,6 @@ class TestRecurrentNetwork:
         assert model_dict['model']['features_per_month'] == features_per_month
         assert model_dict['hidden_size'] == hidden_size
         assert model_dict['rnn_dropout'] == rnn_dropout
-        assert model_dict['dense_dropout'] == dense_dropout
         assert model_dict['dense_features'] == dense_features
         assert model_dict['include_pred_month'] == include_pred_month
         assert model_dict['include_latlons'] == include_latlons
@@ -76,10 +74,9 @@ class TestRecurrentNetwork:
         dense_features = [10]
         hidden_size = 128
         rnn_dropout = 0.25
-        dense_dropout = 0.25
 
         model = RecurrentNetwork(hidden_size=hidden_size, dense_features=dense_features,
-                                 rnn_dropout=rnn_dropout, dense_dropout=dense_dropout,
+                                 rnn_dropout=rnn_dropout,
                                  data_folder=tmp_path, include_monthly_means=True)
         model.train()
